90 % 구현 완료

+++ spring-jpa-sigle-proejct/src/main/java/com/example/order/repository/OrderRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.order.model.Order;
 
-public interface OrderRepository extends JpaRepository<Order,Integer>{
-
+public interface OrderRepository extends JpaRepository<Order,Long>{
+	
+//	@Query("select o from Order o where o.orderer.userName = :userName")
+//	List<Order> findByUserName(@Param("userName") String userName,Pageable pageable);
 }
